% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing variables across the \emph{entire} dataset using MICE.}
\usage{
imputeMissing(
  data,
  flagLevels = c(0, 1),
  miceMeth = "pmm",
  miceSeed = NA,
  miceM = 5,
  miceMaxit = 50
)
}
\arguments{
\item{data}{The data to extract numeric variables from and subsequently truncate.}

\item{flagLevels}{The values of the "_Imputed_Flag" variable in the format c(unmodified, was NA). Can be either numeric or character.}

\item{meth}{The method to be passed to the mice() function.}
}
\value{
Dataframe with added "_Imputed_Flag" variable and changed values.
}
\description{
This accepts the \emph{entire} dataset and returns the \emph{entire} dataset with new variables.
This does \strong{not} preserve original values, as it only changes NAs, and those are encoded in "_Imputed_Flag" variables. No info is lost.
}
