% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeSource.R
\name{safeSource}
\alias{safeSource}
\title{Safely source a file.}
\usage{
safeSource(file, packages)
}
\arguments{
\item{file}{The file to source.}

\item{packages}{The packages that must be installed and/or loaded to safely source the file.}
}
\value{
Nothing.
}
\description{
This function sources a file while installing and/or loading the specified packages,
all in one function. This reduces the need for multiple install.packages or library
calls during impromptu work. No extra error handling is included, so normal package
and source errors will be thrown if applicable.
}
\details{
This was my first ever published function.
}
