% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{summaryFunctions}
\alias{summaryFunctions}
\title{Run summary functions on the supplied data with no additional arguments.
Values returned as a list.}
\usage{
summaryFunctions(data, funcs = c("mean", "median", "range", "IQR", "fivenum"))
}
\arguments{
\item{data}{The data to run the functions on.}

\item{funcs}{The functions to execute, formatted as a character vector of function names.}
}
\value{
Named list of summary function return values.
}
\description{
Run summary functions on the supplied data with no additional arguments.
Values returned as a list.
}
