% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateOutliers.numeric.R
\name{truncateOutliers.numeric}
\alias{truncateOutliers.numeric}
\title{Truncate outliers in entire dataset to 1.5*IQR standard.}
\usage{
truncateOutliers.numeric(data, flagLevels = c(0, 1, 2))
}
\arguments{
\item{data}{The data to extract numeric variables from and subsequently truncate.}

\item{flagLevels}{The values of the "_OoB_Flag" variable in the format c(unmodified, too low, too high). Can be either numeric or character.}
}
\value{
Dataframe with added "_OoB_Flag" and "_Truncated" variables.
}
\description{
This accepts the \emph{entire} dataset and returns the \emph{entire} dataset with new variables.
}
