% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateOutliers.R
\name{truncateOutliers}
\alias{truncateOutliers}
\title{Truncate outliers in entire dataset to 1.5*IQR standard.
This accepts the \emph{entire} dataset and returns the \emph{entire} dataset with new variables.}
\usage{
truncateOutliers(data, flagLevels = c(0, 1, 2))
}
\arguments{
\item{data}{The data to extract numeric variables from and subsequently truncate.}

\item{flagLevels}{The values of the "_OoB_Flag"s in the format c(unmodified, too low, too high). Can be either numeric or character.}
}
\value{
Dataframe with added "_OoB_Flag" and "_Truncated" variables.
}
\description{
Truncate outliers in entire dataset to 1.5*IQR standard.
This accepts the \emph{entire} dataset and returns the \emph{entire} dataset with new variables.
}
